Passando valores para float

+++ global.h
@@ -8,10 +8,10 @@ typedef struct{
 }Auxiliar_de_criacao_de_obj; // criar personagens inimigos TROCAR DEPOIS PARA OBJETOS
 
 typedef struct{
-    int posX;
-    int posY;
-    int altura;
-    int largura;
+    float posX;
+    float posY;
+    float altura;
+    float largura;
     int qtdvidas;
 }Jogador;
 
@@ -19,18 +19,18 @@ typedef struct{
     //int velocidade_do_projetil; // settar as padrões 
     //int quantidade_de_projeteis;
     //int tamanho_do_projetil; 
-    int posX;
-    int posY;
-    int largura;
-    int altura; 
+    float posX;
+    float posY;
+    float largura;
+    float altura; 
     bool atirar;   
 }Arma;
 
 typedef struct{
-    int posX;
-    int posY;
-    int largura;
-    int altura;
+    float posX;
+    float posY;
+    float largura;
+    float altura;
     bool vivo;
     int qtdvidas;
     bool dash;
